requirements.txt
# EasyTSAD项目依赖包
# 生成时间: 2025-06-01 14:26:22
# Python版本: 3.9.22
# 包数量: 63

Deprecated==1.2.18
EasyTSAD==0.3.0.2
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
axial_positional_embedding==0.2.1
bytes==0.0.1.dev4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
einops==0.4.0
filelock==3.13.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
idna==3.10
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
lightning-utilities==0.14.3
llvmlite==0.43.0
local-attention==1.4.4
matplotlib==3.7.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.2.1
numba==0.60.0
numpy==1.23.5
packaging==25.0
pandas==1.5.3
patool==1.12
patsy==1.0.1
pillow==11.2.1
product_key_memory==0.1.10
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2025.2
reformer-pytorch==1.4.4
scikit-learn==1.2.2
scipy==1.10.1
six==1.17.0
sktime==0.16.1
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchmetrics==1.7.1
torchvision==0.20.1+cu121
tqdm==4.64.1
typing_extensions==4.13.2
tzdata==2025.2
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
